// cruzroja-frontend1/src/types/usertType.ts
export type SkillQuota = { id: string; qty: number };
export type CreateEventForm = {
  ambit: string;
  classification: string;
  applyDecreet: boolean;
  marcActivity: string;
  startDate: string;
  endDate: string;
  name: string;
  description: string;
  department: string;
  city: string;
  streetAddress: string;
  sectionalId: string;
  groupId: string;
  attendant: string;
  capacity: number;
  isVirtual: boolean;
  isPrivate: boolean;
  isAdult: boolean;
  isEmergency: boolean;
  participants?: string[];
  requiresSkills?: boolean;
  skillsQuotasList?: SkillQuota[];
};